Added Friend Activity tab to Social Page and made small tweaks to social page
I decided to combine the currently reading section and friend activity sections since there was a lot of extra space on the page with just friends activity and I think that they both fall under "activity". I made a space for the currently reading data to be loaded in activity.html once its created.